+++ src/crypto/hash/cuda/keccak.cuh
@@ -13,4 +13,4 @@
 
 #pragma once
 #include "config.h"
-void mcm_cuda_keccak_hash_batch(BYTE * in, WORD inlen, BYTE * out, WORD n_outbit, WORD n_batch);
+void mcm_cuda_keccak_hash_batch(uint8_t * in, uint32_t inlen, uint8_t * out, uint32_t n_outbit, uint32_t n_batch);
