+++ src/crypto/hash/cuda/md2.cuh
@@ -18,4 +18,4 @@
 
 #pragma once
 #include "config.h"
-void mcm_cuda_md2_hash_batch(BYTE* in, WORD inlen, BYTE* out, WORD n_batch);
+void mcm_cuda_md2_hash_batch(uint8_t* in, uint32_t inlen, uint8_t* out, uint32_t n_batch);
